added sample for React18
