+++ sandstone/react-18/src/App/App.module.less
@@ -1,3 +1,11 @@
 .app {
 	margin-bottom: 18px;
+
+	.tabLayout {
+		overflow: hidden;
+	}
+}
+
+.route {
+	height: fit-content;
 }
