+++ sandstone/react-18/src/views/Batching/package.json
@@ -1,0 +1,3 @@
+{
+	"main": "Batching.js"
+}
