+++ sandstone/react-18/src/views/Suspense/package.json
@@ -1,0 +1,3 @@
+{
+	"main": "Suspense.js"
+}
