+++ sandstone/react-18/src/views/UseTransition/package.json
@@ -1,0 +1,3 @@
+{
+	"main": "UseTransition.js"
+}
